fix(data): move ephemeral_store to infrastructure layer, bump DDD ratchet
- Move ephemeral_store.ts from domain/data to infrastructure/persistence
  to fix DDD layer rule violation (domain must not import infrastructure)
- Bump KNOWN_DOMAIN_INFRA_VIOLATIONS to 26 for CompositeDataQueryService
  extending DataQueryService (inherits same CatalogStore import)
